Rendering fix

+++ src/Main executable/Fastdraw.cpp
@@ -366,7 +366,7 @@ void ShowRLC( int x, int y, void* PicPtr )
 				//clipping right code
 				//cmp		bx, cx
 				//jge		ok2
-				if (TmpBX >= TmpCX) goto ok2;
+				if (static_cast<short>(TmpBX) >= static_cast<short>(TmpCX)) goto ok2;
 				//clipping
 				//cmp		bx, 0
 				//jle		ntd2
@@ -778,7 +778,7 @@ void ShowRLCi( int x, int y, void* PicPtr )
 				//clipping right code
 				//cmp		bx, cx
 				//jge		ok2
-				if (TmpBX >= TmpCX) goto ok2;
+				if (static_cast<short>(TmpBX) >= static_cast<short>(TmpCX)) goto ok2;
 				//clipping
 				//cmp		bx, 0
 				//jle		ntd2
